Refactor 0.1.0
Added: _node, _tree, metrics, preprocessing, utils
